+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+2.4.0
+-----
+
+ * [BC BREAK] made descriptors use output instead of returning a string
+
 2.3.0
 -----